This submission implements the correction for CsMap ticket #4/
This defect was caused by a coding error made sometime in the 2000 time period.  
As this defect affects the numerical results, the 3PARAMETER transformation was 
deprecated, but remains in place, and a new transformation, named GEOCENTRIC has 
been developed.  The new transformation is the same as the deprecated 3PARAMETER 
transformation with the coding error corrected.


git-svn-id: https://svn.osgeo.org/metacrs/csmap/trunk/CsMapDev@27 4e78687f-474d-0410-85f9-8d5e500ac6b2