Submitted to CS-MAP SVN per Trac #16.
Corrected several small memory leaks in the CS-MAP library.  Essentially, the 
datum conversion close function for several datum transformations (Mew Zealand, 
Australian, Germany, and Spain) set the pointer to the conversion shell to the 
null pointer without deleting the conversion shell.  Thus, conversions in any of 
these areas would introduce a memory leak each time they are used,

The intent of the design of the close function is to release basic resources, 
but leave the shell in place in case it needs to be reused; the shell being a 
time consuming object to construct.  Also, this design includes some legacy 
aspects to it.  The shell itself does not occupy much memory, nor does it have 
any file descriptors open.  Thus, the design intent of the close function is to 
leave the shell in place, but release the large memory items and all file 
descriptors.  (Way back when, there was a limit on the number of file 
descriptors you could have open at any given time.

This submission corrects this coding error in the five places it occurs, and 
leave the original design intact, even though for modern systems it is probably 
not necessary.

This submission also corrects some issues in the console test program which 
prevented the memory leak detection feature from working under the newer 
Microsoft compilers.  The memory leak detection feature can now be activated by 
declaring the __HEAP__ preprocessor macro on the build of the console test 
program.  Two memory leaks in the test program itself were also corrected.

NOTE: There is an intentional memory leak encoded into the console test program 
specifically for the purpose of verifying that the Microsoft memory leak 
detection feature is active and properly working (so we'll know in the future if 
it stops working).


git-svn-id: https://svn.osgeo.org/metacrs/csmap/trunk/CsMapDev@1484 4e78687f-474d-0410-85f9-8d5e500ac6b2