CS-Map RFC 1 - Creation of a GDC file to handle French grid files
Implemented the Geodetic Data Catalog feature for the French datum conversions 
between RGF93 and NTF.  This enables French localities to use locally generated 
datum shift files in the NTv2 format in addition to the standard nationally 
generated "gr3df97a.txt" data file.  The grid cell sizes in the respective data 
files causes CS-MAP to always use the local files in those cases where the 
coordinate to be converted is covered by the local file.

The new Geodetic Data Catalog file is named "Rgf93ToNtf.gdc" and is required to 
reside in the same directory as the coordinate system dictionary files.  In the 
distribution, the "gr3df97a.txt" file is moved to the France sub-directory, and 
local grid files are expected to be maintained there as well.

TESTING PERFORMED:
Eleven test cases were obtained from a French locality as used to test the 
implementation.  Prior to the effort, a list of 60 randomly generated 
coordinates using only the national gr3df97a.txt file shift was captured and 
added to the internal test program to verify that there were no regressions.  
These test passed on the Windows 32 and Linux platforms.

Whenever a coordinate is converted, all datum shift sources listed in the 
Geodetic Data Catalog file are examined and those sources whose extents of 
coverage include the point to be converted are noted.  The specific source with 
the densest grid (i.e. smallest grid cell size) is chosen.  As the local files 
are much denser than the national file, the local grid file is always chosen 
when appropriate.

The Geodetic Data Catalog file is named Rgf93ToNtf.gdc as this is the direction 
of the shift data contained in the nation gr3df97a.txt file.  Note that the 
local grid files tend to be in the opposite direction; i.e. the grid shift data 
contained in these files contains the latitude and longitude shift from NTF to 
RGF93.  This added some complexity to the implementation of this feature.

Upon this submission, the implementation does not match the eleven test points 
provided to the degree that we would normally expect.  I suspect that the 
problem is that our implementation of the Lambert Tangential does not produce 
the exact same numbers as that of the French Circe 2000 program.  I've done the 
same test using the Lambert Conformal Conic, Single Standard Parallel projection 
(technically the same projection, but a different implementation) and obtained 
the exact same results as our implementation of the LMTAN.  Also, the creep test 
on our implementation of the LMTAN indicates a creep of less than 2 millimeters 
per 1,000 round trips.  So, this discrepancy remains unresolved.  The 
discrepancy as of this submission averages about 2 millimeters, but gets as high 
as 8 millimeters.  A discrepancy of less than 2 millimeters would be normally 
expected.

git-svn-id: https://svn.osgeo.org/metacrs/csmap/trunk/CsMapDev@1499 4e78687f-474d-0410-85f9-8d5e500ac6b2