This submission corrects Trac tickets 33 and 31; the corrections to these two defects share a common code module cs_wellknowntext.cpp
With regard to ticket 31:  The projection mathematics of conic projections which accept two standard parallels do not care about the order of the parallels.  CS-MAP has always put the northernmost parallel in prj_prm1 and the southernmost parallel in the prj_prm2.  This works fine until you want to compare definitions, for example, derived from WKT where the order of the parallels might be switched.  The change here is to: a> compare the standard parallels in a order independent manner, and b> when reading WKT, always insure that the northernmost parallel ends up in prj_prm1.

With regard to ticket number 33:  WHile ESRI WKT always has the prime meridian value in units of degrees, Oracle (both 10 and 9) and GeoTools output the prime meridian in the units of the host GEOGCS coordinate system.  Thus, CS-MAP has been readingt these things wrong for man years.  Not a big consequence, as the following must be in effect for this defect to be observable:

1> Prime meridian must be something other than Greenwich
2> GEOGCS system unit must other than degrees (e. g. grads)
3> The flavor of the WKT being processed must be Oracle9, Oracle, or GeoTools.

I have no information about what other flavors do in this case. so those have been left alone until such information is available.

git-svn-id: https://svn.osgeo.org/metacrs/csmap/trunk/CsMapDev@1556 4e78687f-474d-0410-85f9-8d5e500ac6b2