This submission addresses Trac ticket 36.
The original CS-MAP console test program was written in C, as was the entire CS-MAP library at that time.  As more and more C++ has crept into the CS-MAP library, having the console test program as a C program had become quite inconvenient.  This submission submits a version of the CS-MAP Console Test program which has been modified so that it all complies as C++, making it rather easy to call, exercise, and test features of the library which are written in C++ as well as those written in C.

The primary driving factor behind this submission was the desire to access our new TcsEpsgDataSetV6 object in our console test program.  There are two new tests:

CStestM.cpp  ->  Implements the Name Mapper audit test, where the names and ID's maintained in the Name Mapper database are compared to external data sources for accuracy and concurrency.  Currently, only the EPSG external database is used in the testing process.

CStestN.cpp  -> Implements the CS-MAP Dictionary and EPSG Dataset comparison test.  For each definition in a EPSG Dataset for which the Name Mapper database contains a mapping, the EPSG version of the definition is compared with the CS-MAP version.  This verifies that the Name Mapper maps to the correct definition, and also flushes out any discrepancies in the definitions which might exist.

Finally, To implement TestN in a rational manner, a more discerning set of compariaon functions was necessary.  The following three functions were added and are used solely in TestN of the C++ version of the console test program.

CS_elDefCmpEx ()  -> A new function which compares ellipsoid definitions returning a double value which indicates the degree to which the compared ellipsoids are the same.

CS_dtDefCmpEx ()  -> A new function which compares datum definitions returning a double value which indicates the degree to which the compared datum definitions are the same.

CS_csDefCmpEx ()  -> A new function which compares coordinate system definitions returning a double value which indicates the degree to which the compared coordinate systems are the same.



git-svn-id: https://svn.osgeo.org/metacrs/csmap/trunk/CsMapDev@1574 4e78687f-474d-0410-85f9-8d5e500ac6b2