Trac #54 --> Fix ordering of the NameMapper.csv file.
The order of entries in the NameMapper.csv file is immaterial to the operation of the NameMapper itself as each record in the .csv file is read into a std::set object which (essentially) reorders the records upon construction.  In the original implementation, the function which sorts the NameMapper.csv file failed to include a complete unique key.  Thus, each time the NameMapper.csv file is generated, there are typically a 20 or so instances of the records being randomly reordered.  This does not affect operation, but it makes for a rather difficult to read "diff" report.

This submission augments the key which is used to sort the NameMapper.csv file so that a "diff" reports only the real material changes to the NameMapper.csv file.

The actual data content of the NameMapper.csv file submitted here is identical to that of the previous revision; only the order of the records has changed.  Thus, this revision o fthe NameMapper.csv file serves as boundary between the previously poorly sorted version and the new correctly corted version.


git-svn-id: https://svn.osgeo.org/metacrs/csmap/trunk/CsMapDev@1927 4e78687f-474d-0410-85f9-8d5e500ac6b2