Trac Tickets: 75, 77, 78, 79, 80, 81, 82, and 83.
As part of the RFC 2 implementation, a rigorous regression test was devised.  This procedure involved the production of some 64,000+ test points converting between various geodetic datums using the previous version of CS_MAP.  We cared not about how accurate or precise those results were, just that they were produced with the previous release.  Yet another application was then written to do the same calculations using the RFC 2 implementation.  The results were compared with a tolerance of 0.05 millimeters (1.0e-09 in terms of latitude and longitude in degrees).

This test originally produced thousands of failures.  The following corrections resolve all of these issues.  As the changes for each Trac ticket item are scattered avross several functions, and time to get all corrections done was very short, these changes are all submitted as a unit.


git-svn-id: https://svn.osgeo.org/metacrs/csmap/trunk/CsMapDev@1943 4e78687f-474d-0410-85f9-8d5e500ac6b2