Trac # 90.  This submission contains additional corrections for the issues of #90.  Essentially, it corrects issues related to NameMapping and the Coordinate SYstem Dictionary changes of the last 12 months including the major changes with regard to the HARN systems.
For some reason related to external data source standards, we were asked to add 
the CRS:84 definition to the dictionaries; this definition being identical to 
LL84, just having this unique name.  As a result, the GEOGCS portion of CRS WKT 
renditions specified CTS:84 as the GEOGCS portion of the WKT.  This was 
considered undesirable and the CSllFromDt function of the CS_csprm.c module was 
hacked to simply return "LL84" whenever the datum provided was 'WGS84".  This 
corrected this issue,

The remainder of the changes were made to the NameMapper file.  There are 
several hundred changes here which can be organized into the following 
categories:

1> Several duplicate name definitions were removed.  These duplicates originated 
from the fact that at the time the major portion of the Name Mapping effort was 
made, ESRI WKT sources had several definitions which EPSG did not support.  When 
these definitions were added to EPSG over time, we ended up with duplicate 
definitions in the Name Mapper which produced erroneous operation of the WKT 
processor.

2> In Tavaura, we under took an effort to eliminate duplicate definitions.  That 
is, deprecate (move to the Obsolete category) definitions which were considered 
the less desirable of a pair of duplicate definitions.  The csApiTest 
application, which checks for WKT regressions was not run prior to that release. 
Several issues with Name Mapper data caused by this removal of duplicates were 
corrected.  Specifically, the name mapping of the deprecated systems was 
adjusted to reflect the fact that they were deprecated.

3> Two erroneous entries in the category.asc file were removed.

4> A bogus Oracle SRID reference in the coordsys.asc file was corrected.

5> Corrected several regressions caused by EPSG deprecations which were included 
in the Tavaura release.  Essentially, EPSG name changes which were incorporated 
into Tavaura produced regressions which were not detected in Tavaura as the 
csApiTest regression test was not performed due to emergency defect corrections 
required to be made to the MGRS grid feature.

6> Located and corrected a mismatch of systems.  (US48)



git-svn-id: https://svn.osgeo.org/metacrs/csmap/trunk/CsMapDev@1955 4e78687f-474d-0410-85f9-8d5e500ac6b2