Fixing ticket 119 (Transformations going from NAD27 to new HARN/... datums are not using HARN grid files)
The main problem fixed by this submission is that the HARN/... datums where *NOT* using the HARN grid files when transforming between NAD27 and HARN.

This was due to missing path definitions.
When transforming between NAD27 and HARN, the transformation bridge built on-the-fly did not pick up the individual NAD83_TO_HARN/... transformations but used the NAD83_TO_WGS84 instead. 

This submission contains 2 types of fixes:

1) create 50+ new paths of type [HARN/XX_to_WGS84] analog to the existing [HPGN_to_WGS84]
These guys look as follows:

GP_NAME: HARN/AL_to_WGS84
	  SRC_DTM: HARN/AL
	  TRG_DTM: WGS84
	  DESC_NM: High Accuracy Regional Network (aka HPGN) for region AL
	   SOURCE: Derived by Mentor Software from US NGS NADCON 2.10
	    XFORM: NAD83_to_HARN/AL,INV
	    XFORM: NAD83_to_WGS84,FWD

2) Move the HPGN-based systems out of the LEGACY group.
As stated in the ticket, the HPGN-based systems, like [VAHP-SF], are not incorrect in terms of that parameters are bad or something like that - there merely exist alternatives, like [HARN/VA.VA-SF], which might be more appropriate to use in a given context.

The systems that have been moved out of the LEGACY group have been distributed to the appropriate [USA, ...] and [UTM, HPGN Datum] categories.

Changes have been unit-tested + reviewed by Hugues.

git-svn-id: https://svn.osgeo.org/metacrs/csmap/trunk/CsMapDev@2143 4e78687f-474d-0410-85f9-8d5e500ac6b2