Fix ticket 149 "Namemapper.csv incorrectly maps "foot" to "US Foot" for Oracle 10 and 11"
When CS-Map parses a WKT and no flavor was given, it gets the several WKT elements and checks the values 
against the namemapper.csv file.

For each relevant WKT element (e.g. projection, units for both the GEOCS and 
PROJCS, parameters), CS-Map builds a bit-mask of the assumed flavors and then 
strictly AND's them to find the assumed target flavor.

This failed for several Oracle WKTs because the namemapper only contained an entry for 
"U.S. Foot" for Oracle 9 but not for Oracle 10. This did result in the flavor 
detection to fail, because CS-Map found the entire WKT to be in Oracle format, 
i.e. 10 or 11, but just the unit in Oracle 9 what did result in the flavor 
detection bitmask to become 0 again.

With that error being returned to the MgGeometry.dll, the code then always re-starts the parsing with a specific 
flavor, first OGC, then ESRI and so forth until a match is found. In some cases, 
CS-Map then really finds a "match" - in some cases when using the ESRI flavor; 
for the [NAD83 / Texas Central (ftUS)], SRID 2277, the WKT got parsed as HARN-referenced CS, what's not correct at all.

The fix is to correct the namemapper.csv so it contains the correct entries for 
"foot" and "U.S. Foot". I corrected the following entries:

//for international foot
9002,6,Oracle,0,"Foot (International)",0,1,0,0,,
9002,6,Oracle,0,"foot",0,1,0,0,,


//for U.S. Foot
9003,6,Oracle,0,"U.S. Foot",0,0,0,0,,

Reviewed by Hugues.

git-svn-id: https://svn.osgeo.org/metacrs/csmap/trunk/CsMapDev@2279 4e78687f-474d-0410-85f9-8d5e500ac6b2