Partial resolution of Trac Ticket 112.  This submission adds about 700 (after conflict removal) useful range definitions to CRS definitions in the coordinate system dictionary.  This was accomplished programmatically by pairing CRS definitions without a useful range with "similar" CRS definitions which a useful range specification. Similar, in this case, means essentially the same coordinate system, but with different datum and/or linear units.  Differences in these two items have no significant affect on the useful range.  A copy of the the table used was attached to the Trac Ticket (#112).
This submission also include some significant improvements to the Useful Range Report and CsdToCsvCS features of the Console Utilities project.  These were helpful in arriving at the CRS pair list which was used in this process.

Ticket 112 is still not fully resolved as there remain 668 non-Legacy definitions without a useful range specification.  The improved Useful Range Report suggests that there may be as many as 150 definitions where the CS-MAP useful range is significantly different from that which appears in the EPSG dataset.


git-svn-id: https://svn.osgeo.org/metacrs/csmap/trunk/CsMapDev@2565 4e78687f-474d-0410-85f9-8d5e500ac6b2