Still Working Trac #179.  This submission h as the bulk of the correction for the subject Trac Ticket.  Most all NSRS 2011 based systems now properly have the correct EPSG code in both the Coordinate System Dictionary and the NameMapper.  The NameMapper has correct EPSG and ESRI names along with correct ESRI codes.
There remain about 35 system which will have to be done manually due to a screw up in the initial NSRS 2007/2011 project.  Essentially, the geographic systems related to these two new datums were ignored, as were the state plane coordinate systems of Alaska.

Alaska represents a special problem which requires more research.  All of the data files involved in the NSRS 2007 conversions convert from NAD83(HARN) to NSRS 2007.  I am unaware of any conversion from NAD83 to HARN for Alaska.  Once reseach of this issue is complete, we'll know exactly what to do about Alask with regard to NSRS 2007 and 2011.

Most all of the changes included in this submission were accomplished programmatically, the code used in included in the submission.  Its not very nice code, but it is only intended to be used once.

This submission also includes code for a utility to convert the dump of SRD's from the typical Oracle format to the .CSV format we like to use.

git-svn-id: https://svn.osgeo.org/metacrs/csmap/trunk/CsMapDev@2573 4e78687f-474d-0410-85f9-8d5e500ac6b2