Update CS-MAP to Release 8.05 of EPSG Parameter Dataset.  Doe not include new definitions.  This submission updates names, includes deprecations, and minor changes.  The effort included the comparison of various names with the WKT catalogs for ESRI and Oracle.
There remain several hundred new definitions in EPSG, ESRI, and to a lesser extent Oracle, which have not been added to the CS-MAP dictionaries.  This will take a significant amount of time and effort and is not included in this submission.

The submission has been tested on four versions of VC++ (9.0, 10.0 11.0, & 12.0) and Linux using CentOS and "gcc" version 4.7.

The CsMapConsoleUtility code submitted was used at various stages of the update to automate the update process as much as possible to reduce the probability of typographical errors.  This code should not be considered to affect the performance or functioning of the library itself.

There exists on this authors system a catalog of WKT strings and associated EPSG identification codes.  This is not submitted as the data is not known to be in the public domain.


git-svn-id: https://svn.osgeo.org/metacrs/csmap/trunk/CsMapDev@2584 4e78687f-474d-0410-85f9-8d5e500ac6b2