Working Ticket #187.
This submission adds the production of a Windows DLL for each of the five Windows compilers supported.  The DLL's are named CsMapDll.dll and are written to the appropriate bin??? directory with the CsMapDev directory.  The related .lib and.exp files as relates to the DLL are written to the same directory.

It is the specific intent of this development that the DLL be as generic as possible, thus the appropriate CRT is statically linked into each DLL module.  The only external references required by the DLL (to the best of my ability to determine this) is kernel.dll and oldnames.dll.  Thus, the DLL is expected to load and properly function in any VBA environment.

Two header files are included in the 'Include' directory to help users use the DLL.  'cs_mapDll.h' will be useful to 'C' and 'C++' programmers.  'Header.vba' will be useful for Visual Basic programmers (and an assist for other fourth generation language programmers).  Additionally, a 'DllReadMe.txt' file is included in all five VC??? folders with comments pertaining to the DLL.  Currently, all five reademe files are identical; perhaps there will be differences which will need to be recorded later on.

With one exception, no significant code changes were required to produce the DLL.  Most all of the changes in the .c and .h modules included in this submission are simply changes to the EXP_LVL? defines and their use.  As has been the case for decades, all of these defines are set to the null string when compiling the ordinary object library.  Thus, there should be no change in CS-MAP functionality in the normal case.  The one exception referred to above is CS_guiApi.c.  The code for two (or is it three) functions was adjusted to work as documented.  These functions were originally devised for the MFC dictionary functions which have been deprecated for some time now.

A new class of functions have been introduced with this submission, all of which emulate an existing function, and have a 'VB' appended to the name of the existing function.  Essentially this class of functions supports use of the DLL in the VBA environment without resorting to dealing with the BStr problem.  Avoiding the manufacture, allocation, and release of BStr's was chosen as the approach implemented is much more reliable and less error prone in the wide variety of environments in which the DLL could be used.  Essentially, this class of functions perform in exactly the same manner as the original, but the string returned in the references character array is not null terminated and the result is always space filled to the extent of the character array.  These functions are all defined in the new CS_vbApi.c source module.



git-svn-id: https://svn.osgeo.org/metacrs/csmap/trunk/CsMapDev@2665 4e78687f-474d-0410-85f9-8d5e500ac6b2